Added session to VRPresentationFrame
This should be relatively uncontroversial, but for the sake of
explanation:

When a page is managing multiple sessions (for example a non-exclusive
magic window session that then gives way to an exclusive session of the
same content and then back to magic window when it’s done) you will
frequently want to use the same render loop callback for both sessions
because you’re rendering the same content. By allowing developers to
query the session that generated the frame from the VRPresentationFrame
that’s passed into the render loop it’s much easier to handle those
kind of scenarios and means less fragile state tracking in general.